Adds metrics for how many permissions are requested, and how many queued.
These metrics are to help determine whether or not the current mechanism
of deciding when to show the bubble for coalesced permissions is reasonable.

In other words, are there too many queued (waiting) permissions that will
appear when the visible bubble is dismissed?

BUG=368737

Review URL: https://codereview.chromium.org/260933002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bc80973ac4906e1a04851b086f0a96863d0fec78