Added UMA histograms for tracking desktop vs. mobile site page loads.
Added two histograms. One records the type and details of a viewport
specification on the page. The second is recorded in the case of a
fixed-width viewport meta tag. It records the percentage zoom that
is required to fully see the page on the device's screen.

Associated Blink patch: https://codereview.chromium.org/216183004/

BUG=329697

Review URL: https://codereview.chromium.org/212553006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 377d991573feb860f8cf37bbaaaae47a6c92e6c7