Fix compile error if both inline and reference enums are used in a type.
The ToString() overloading function of inline and reference enums
will conflict because they exist in both global namespace and local
classes. This is a workaround to generate EnumNameToString() and use
it in ToValue() of the type. ToString() is still generated for
compatibility because some other code still use it.

R=kalman@chromium.org
BUG=371042

Review URL: https://codereview.chromium.org/270523005

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 29189e1dce5e4cee4d29e5e8a86494f60f597fa9