[Telemetry] Remove power benchmark from JSGameBench and Spaceport
Measuring energy usage in JSGameBench, Spaceport and Scirra is meaningless since the amount of work/wall time for these tests isn't meant to be stable or related to the resulting score.

Note: Scirra doesn't currently measure energy usage in Telemetry so there was nothing to remove and Spaceport is currently disabled on Mac (the only platform that measures energy usage on the waterfall at the time of this writing).

BUG=None

Review URL: https://codereview.chromium.org/270223009

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 395d2a0a88c40cfb45f604fcf12e9f968cd0b644