Do not mark alternate-protocol broken unless the main job succeeds.
Currently, if both the main and alternate jobs fail, we mark alternate-protocol as broken. However, that's not really "fair" to the alternate-protocol. We should only mark it as broken in the alternate-protocol jobs fails and the main job succeeds.

This is a bit tricky because the two jobs may complete in any order.

Review URL: https://codereview.chromium.org/245663003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 00c159ff2648341770f89d51269bf098f2a6dce4