Disk cache: Remove stats_histogram.
StatsHistograms is the only "external" subclass of base::Histogram and as such
it often requires extra considerations when modifying the base class. On the
other hand, almost the same data can be gathered by using a regular log
histogram with a specific range.

This CL does that, generating a 75 bucket histogram to map the 27 buckets
from the original data. Most of the buckets will be empty, and a couple of
buckets will be merged on the new histogram, but the compromise looks
good enough.

BUG=377936

Review URL: https://codereview.chromium.org/295913014

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a3cfbfaab843a48399ffb31a73f98366e88417f0