Disable ActionRunnerTest.testTapElement on Win Debug.
This test is crashing on the Win (dbg) bot on the
chromium.webkit waterfall; the test was added in
https://codereview.chromium.org/323833003/ and
has never run successfully, so this may be exposing
a real bug. Disable it temporarily to green up
the tree while we investigate.

TBR=tonyg@chromium.org
BUG=384181

Review URL: https://codereview.chromium.org/331773003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: daec42a5332a7e05a186a1c794c9cc86057b6024