Fix leak WebURLLoaderImpl::Context leak on cancel.
The Context was implicitly owned by the ResourceDispatcher as
well as the WebURLLoaderImpl, and was leaked on cancel because the
ResourceDispatcher would never call back into the context like it
expected.  This CL removes that implicit ownership, and instead
just keeps the Context from being deleted while it's on the stack,
and still has work to do.

BUG=328092

Review URL: https://codereview.chromium.org/342203002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a491fd9412b763f20244acb4aea17312a89f64da