[PowerProfiler] pySerial module serial.tools.list_ports for linux should include the product information as description With monsoon power monitor connected, while using serial.tools.list_ports module to see available ports, the description field returned is always 'ttyACM0'. Whereas the same API on windows and mac, returns 'Mobile Device Power Monitor'.
In this CL, we try to retrive the product information from the /product file associated
with this device. If we are unable to retrive, we simply return the device name itself
as is the present case.

BUG=382315
NOTRY=true

Review URL: https://codereview.chromium.org/325663002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0ee159723fa8ceffde9c021d37ee47363f5190c5