Remove Net.URLRequest_SetReferrer_IsEmptyOrValid histogram.
According to internal stats, this was the top 5 histogram
that's taking up space in UMA and it was never documented
in histograms.xml, which means no one has ever looked at
its data. Delete it so that it stop taking resources without
providing value.

Document it in histograms.xml so that previous data can
still be viewed on the dashboard, if needed.

BUG=387734

Review URL: https://codereview.chromium.org/331343003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d851d064ee06e6835b7ef153ec4afe5c3aeb77ee