Add UMA for the new generic permisison class
In order to track UMA accurately we only need to track it when there is an actual prompt for permission, not when the permission was saved for the domain. I have re-purposed the PermissionDecided method for this which was essentially doing nothing before.

BUG=392145
TBR=fgorski,tommi

Review URL: https://codereview.chromium.org/371933002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3ce59a2f2be2b40e38f8e2ad75ba6f0476993c47