[Telemetry] Disable core-scroll-header-panel which was recently enabled.
It did pass two runs, but is consistently failing now. So I'm not sure
whether this is a regression or flake. Since the test is so new, it seems
safest to just disable this page until it shakes out.

If this isn't a flake, the regression range is r283974-284019.

BUG=394756
TBR=wiltzius@chromium.org
NOTRY=True

Review URL: https://codereview.chromium.org/400823006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 781849a0007462f54b262427eccead697d5d9c6e