Add SkipValue to Telemetry.
This adds logic for representing page skips as values in Telemetry, in the
spirit of how failures are to be handled with regard to the linked bug. The
goal for this CL is just to add the logic, and then once it's safely in the
codebase, we can wire it up and make sure Buildbot is happy with it.

BUG=392901

Review URL: https://codereview.chromium.org/409023002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 125ec622d28e5c347351892cf7f5134957b2dbf9