[usb_gadget p07] HTTP management server.
This simple HTTP server allows remote management of the USB gadget. The
serial number of the device is set to the IP address and port number of
the server so that from a device enumeration an application can then
control the device over the network.

BUG=396682
R=rockot@chromium.org,rpaquay@chromium.org,kalman@chromium.org

Review URL: https://codereview.chromium.org/418773002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b0f98ea914bd4dfdfabb59409e37e0796787f0c1