[usb_gadget p08] Package the USB gadget framework for easy distribution.
This software package will need to be uploaded to the test hardware.
Python can easily run code out of a zip file. The package hash is saved
to verify integrity and as a versioning mechanism to allow tests to
ensure that they are running against a device with the expected software
version.

BUG=396682
R=rockot@chromium.org,rpaquay@chromium.org,kalman@chromium.org

Review URL: https://codereview.chromium.org/414833004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ee8d3b000414b0b3529f0117a706e0f6f6cbd2d2