Close open file handle before moving file when recording.
In record_wpr, we did the following:

  temp_target_wpr_file_path = tempfile.mkstemp()[1]

In python, tempfile.mkstemp() will return a tuple containing the file handle and the file path. Since the script never closes this file handle, it looks like we throw a WindowsError when we try to move the file with the open file handle.

BUG=

Review URL: https://codereview.chromium.org/420893002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: da3ff86aa88a5de1bd1b35d68b2c2558c42154fa