Experience Sampling API implementation
This CL includes the API JSON schema, the implementation of the API functions, and a class for creating and tracking sampling events (to simplify instrumenting UI with the onDisplay event). It also adds a new extension permission for experienceSamplingPrivate, and an empty whitelist.

BUG=384635

Review URL: https://codereview.chromium.org/395313002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 31339076da909609961bb8cdac546fc2646cca4d