Attempt to fix UnixDomainServerSocketTest.AcceptWithForbiddenUser under Valgrind on Mac.
The issue is most likely the assumption the test makes that
running the message loop is enough for a socket message to make it
from the server socket to the client socket.  The fix tries to
read from the client socket instead, and waits for the result.

BUG=397031

Review URL: https://codereview.chromium.org/420823002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 18da15b84d6e7bdf047e3e190efecababa284bac