Formalize the concept of page runs in PageTestResults.
A page run is created on StartTest and finalized on StopTest.
We allow StopTest to discard the entire run (to support
discard_first_run option). We also allow a page run to restart
a new attempt at running the page (via
PageTestResults.WillAttemptPageRun) - this will clear any
existing results for the run.

As a results of this change:
* page_runner.py no longer need to use a hack of creating a
temporary results object.
* When we skip a test, it is now considered a success, so 
the last line of the test will now include the count for skipped
tests (the line that says: [ PASSED ] 1 test. See
gtest_test_results_unittest].

This patch does not yet remove AddSuccess, since it is still
used in GTestTestResults. I plan to move the logic there
to StopTest instead of AddSuccess (similarly for _EmitFailure
and _EmitSkip).

BUG=383639

Review URL: https://codereview.chromium.org/393283008

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 57f2796401f84a3df2a23f2069ad537e82836fdf