Adjust addends rather than targets for RELA configurations.
Relocations with addend use the addend rather that the location pointed
to by the relocation as the target of the relocations.  For RELA
relocations, ignore the target and instead adjust the addend for the
hole being created in the file.  Change unit test data accordingly.

Ensure that sections we resize are never zero-length when we load the
Elf file, and that we never set either to zero-length on packing.

BUG=385553

Review URL: https://codereview.chromium.org/423393010

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 60319add4dc45996150b57148fd1c24763ba7de2