Adding the VerifyModule function (and helpers) to safe browsing.
This function allows us to compare a module in memory with 
its original on disk, and identify any differences (after 
accounting for relocations).  The state of the module 
(MODULE_UNKNOWN, MODULE_UNMODIFIED or MODULE_MODIFIED) is 
returned along with a list of the possibly modified exported 
functions of the module.


BUG=401854

Review URL: https://codereview.chromium.org/406043003

Cr-Original-Commit-Position: refs/heads/master@{#288319}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 848453caed8b55480d71bb9ed7fe8393c4273dfc