[Telemetry] Move failure reasoning into summary.py.
Previously, we were having Summary take a boolean had_failures. Now that
failures are part of Telemetry's value system, it is easy to have Summary
reason about how to summarize values when failures and/or skips are present.
This CL is the first step: it eliminates had_failures from being passed around
and instead computes it directly based on the presence of failures in the
passed list of page-specific values.

Review URL: https://codereview.chromium.org/465063002

Cr-Original-Commit-Position: refs/heads/master@{#289556}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aa3ca1c1d5109aafd1137bda680af54cc4be50c5