[Password Manager] Setup experiment to restrict autofilling of sync credential
By default there is no change in behavior, but autofilling can now be disabled
for the sync credential entirely or just disabled for reauth pages that support
transactional reauth.

Note that this also changes GetSyncUsername() to not return the username if
password sync is disabled if it is possible to determine. This makes GetSyncUsername() a little inconsistent depending on the state of sync setup, but it's important to be as specific as possible when disabling autofilling, since it's a
usability hit.

BUG=386692
R=isherman@chromium.org

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=290030

Review URL: https://codereview.chromium.org/451853003

Cr-Original-Commit-Position: refs/heads/master@{#290057}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 524dd70ed5a850b012a70c9852e7980284360f50