Expand empty directories in GN to "."
When a directory substitution in a toolchain definition expands to the empty
string (i.e. it's the same as the root build directory), expand the name to "."
instead of the empty string to avoid corrupting filenames.

For example, if you specified "{{target_out_dir}}/foo" without this it would
expand to "/foo" for the default toolchain.

This removes the previous special-case hack for this case. That code only
worked when the entire string expanded to the filename (when the directory was
at the beginning of the string). When generating command lines, the directories
will be in the middle of the string and the previous special case didn't catch
them.

R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/497193002

Cr-Original-Commit-Position: refs/heads/master@{#291469}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f5290a91415eb68e701b035968c784c181b31d78