Componentize component_updater: Copy test data over to components/test/data
As part of making component_updater a component, we are now moving over all the tests.

Before we do that, we're copying the test data over. We are deleting it from the original spot in a separate CL.

BUG=371463

Review URL: https://codereview.chromium.org/497803002

Cr-Original-Commit-Position: refs/heads/master@{#291470}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6eef31af4e617ca66acb7a39792506ecde7861ee