[telemetry] Add WinRing0/MSR power monitor.
This is a power monitor that reads the Intel RAPL MSRs to get energy measurements.
It uses a defunct open source DLL and kernel driver, WinRing0, to access the MSRs.
WinRing0 is a little wonky, and we have to install the kernel driver to the Python executable's directory.
Python must be run as Administrator for this to work.

BUG=336558
TEST=tools/telemetry/run_tests winring0

Review URL: https://codereview.chromium.org/536243002

Cr-Original-Commit-Position: refs/heads/master@{#293277}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 10d4df165c5ce09503738fdd0549b5287c00d713