[Telemetry] Fix Chart JSON output typo bug and add unittest
There was an issue with outputting Chart JSON that came down to a typo and was
not caught by unittests. This patch fixes the bug and adds a unittest to make
sure it won't crop up again.

R=nednguyen,sullivan

Review URL: https://codereview.chromium.org/551293003

Cr-Original-Commit-Position: refs/heads/master@{#294239}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 701934e5b0371bea45f09e9942fb5f4ac801ddc4