[Telemetry] Fix handling for two-part per-page value names in Chart JSON.
The perf dashboard expects that for values of the form x.y that are per-page,
the computed summary x.y (i.e., the value having that name but not associated
with any particular Page object) will be displayed but the per-page x.y's will
be ignored.

Previously, the per-page x.y's were being named as x: <pagename> in the
buildbot convention; it was the job of the log processing step to make sure
that these extraneous, ambiguously-named values were ignored. The Chart JSON
output formatter, on the other hand, explicitly guards against adding values
with duplicate names, and so extra checks are necessary to throw away such
extraneous values.

This patch adds handling to the Chart JSON output formatter to address this
issue, and updates the tests to check for correctness in this scenario.

R=sullivan,qyearsley,tonyg

Review URL: https://codereview.chromium.org/551063005

Cr-Original-Commit-Position: refs/heads/master@{#294683}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fd93e29f1f5d03eaf07af9b20f64a568997e874c