GN fixes for group dependent configs.
Adds a better tests for pulling dependent and public configs. These don't actually test the bug, but are important.

The bug was removing the group conditional when pulling dependent configs. Previously the groups deps would get duplicated so we wouldn't want to pull a group's public deps. I didn't write a test for this since it seemed odd to test that there is not a special case for this. I just removed the condition.

Fixes a misspelliung and and an extra newline in the help.

BUG=
R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/581723002

Cr-Original-Commit-Position: refs/heads/master@{#295314}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c4a325a6cef25f7380b9c7cca9178b85849077ae